Really fix linking
Recent commit 56c441d is essentially wrong -- when dynamically linking
a binary with some library, one should NOT explicitly put the dependencies
of the library being linked, as doing that is redundant and leads to
so-called "overlinking" (see [1]).

The problem of linking on recent Ubuntu mentioned in the above commit
was caused by the fact that the libmosaic.so itself is "underlinked":

	# ldd lib/libmosaic.so
	linux-vdso.so.1 =>  (0x00007ffd72932000)
	libc.so.6 => /lib/x86_64-linux-gnu/libc.so.6
(0x00007f1096107000)
	/lib64/ld-linux-x86-64.so.2 (0x00007f10966dd000)

As you can see, libyaml.so is not listed as a dependency, although
it definitely should be. The cause of the problem is the wrong order
of arguments when creating libmosaic.so, plus new gcc toolchain being
more picky about such order.

This patch fixes both issues. Tested on Ubuntu 15.04 and CentOS 7.

[1] https://wiki.mageia.org/en/Overlinking_issues_in_packaging

Signed-off-by: Kir Kolyshkin <kir@openvz.org>