Version the library
A shared library should definitely have a version. It's not just a good
faith or a way to make sure library users request a particular version
of a library, but also a prerequisite for some distributions. As far
as I remember Debian QA have an automatic check that reject software
with a library that is not versioned).

Currently the version is hardcoded in lib/Makefile (as 0.0) -- this can
easily be changed in the future.

Also, this patch:
 - introduces LDLIBS variable (as libraries should not be added to
   LDFLAGS, as libraries must be specified after the objects in case
   --as-needed linker flag is specified).
 - Cleans up "%.o: %.c" target

Signed-off-by: Kir Kolyshkin <kir@openvz.org>