libmosaic.so: hide symbols not in API
In the .so file, we should only expose symbols that are part of the
public API (include/uapi/mosaic.h), and hide all the others.

Currently the only exception to the above is mosaic_parse_config()
that is not part of public API but is used by moctl. Pavel likes
the name "protected" for such objects, so let's call it LIB_PROTECTED.

This change helps to establish clear boundaries preventing bad usage
cases (such as when an internal library function is being called
from the outside). It might also:
 * reduce the library size
 * improves its load time
 * lets the optimizer to produce better code
 * reduces the chance of symbol collision

For more details on the topic, see https://gcc.gnu.org/wiki/Visibility

Signed-off-by: Kir Kolyshkin <kir@openvz.org>