Makefiles: fix building in subdirectories
Without this patch, running 'make' in a subdir is not working for
two reasons:

1. CFLAGS from parent Makefile are not added, as it's not included.

2. If they would be included, they'd still be incorrect.

In general, tricks like this (with pwd etc.) never work, so let's move
back to basics and just add proper -I to CFLAGS in each directory.

Also, move common initial settings (currently just CC and CFLAGS) to
a separate Makefile.inc, and include it where needed.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>