moctl: add a shell wrapper
This is a way to solve a problem when in the same source tree you have
one or more libraries, plus one or more binaries that are dynamically
linked against those libraries, and you want to be able to run those
binaries without installing those libraries first. Pretty natural
thing to do, but requires some minor trickery in the form of a shell
wrapper.

With such a wrapper, you don't have to worry about the above problem,
but there is a "side effect" that needs to be mentioned:

The binary is now not the real binary, but a shell wrapper.

The consequences are, for example:

* you can't use tools like ldd, gdb, strip against the non-installed binary
* you can't install the binary by mere copying -- use "make install"
* installed binary of course still requires installed libraries

Once you know this, you'll be fine and dandy.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>