volume map: initial implementation
This is the initial implementation of the volumeMap feature.
It allows to use regular expressions with \(...\) sub-matches
and \N backreferences to apply some transformations for volume names.

For example, one can write this in .mos file:

volumeMap: ve\([1-9][0-9]*\)\.disk\([0-9a-zA-Z]*\) private/\1/disk/\2

Given the above example, if volume name is specified as "ve123.diskD",
it is transformed to "private/123/disk/D". There can be up to 9 replacements
(\1 to \9), and extended regular expression syntax is to be used, as
described in regex(7) man page. Currently in case there's no match,
an error is returned.

Tests are to be added later.

[v2: move out of layout, made a global not per-driver feature]

Signed-off-by: Kir Kolyshkin <kir@openvz.org>