CHKDUP: replace to CHKVALTO
In case CHKDUP finds there's a duplicate, it returns so we leak val.
In order to not leak it we need to free() it, but CHKDUP doesn't have
val.

So, replace CHKDUP with CHKVALTO which calls CHKVAL first and then
checks for duplicate, freeing the value if needed.

Reported by Converity, CID #15492.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>