moctl: remove 'create' functionality
According to README, "both moctl (the tool) and libmosaic work
on a pre-created storage". According to Pavel, adding create
functionality to moctl was a mistake.

So, ditch it.

Also, this allows us to get rid of the LIB_PROTECTED libmosaic hack
(i.e. exporting functions that supposed to be internal to the library
but are still being used by our tool).

Signed-off-by: Kir Kolyshkin <kir@openvz.org>