logging: don't use 0 as log level
This is weird, but when compiling a recent mosaic with gcc-4.6,
it gives the following warning:

> [root@tpad-ovz1 lib]# make log.o
> gcc -Wall -Werror -Wextra -Wno-unused-parameter -Wno-sign-compare
> -I../include -fPIC -fvisibility=hidden -c -o log.o log.c
> cc1: warnings being treated as errors
> log.c: In function ‘test_log_compile’:
> log.c:23: error: comparison of unsigned expression >= 0 is always true

This happens because
 - gcc uses unsigned for enums that don't contain negative elements
 - loge() macro uses LOG_ERR value of 0

So it's "0 <= d" kind of comparison, where d is unsigned, and the result
is always true. It is totally fine, but gcc-4.6 suspects bad code.

To make it happy again, set LOG_ERR to 1.

This is a fix to commit 78a110f.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>