moctl: fix argument parsing
The idea (as implemented from the very beginning and later described
in commit df652a9) is to allow abbreviated commands, such as using
'mo' instead of 'mount'.

Suprisingly, the implementation allows for enlarged ('mountitformeplease')
but not for abbreviated commands.

To fix, we need to change both the order of checking and the values
returned.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>