Add "have volume" feature v3
There is a need to check if mosaic has a given volume, in other words
if the given volume exists or not. This can be used by e.g. Docker
graph driver.

All the current drivers modified to support the functionality.
Also, fix the "FIXME" in open_volume method for btrfs.

Add "have" command to moctl as a sample interface.

Add some simple tests.

[v2: exists -> have, make the callback required]
[v3: add flags argument, same meaning as for open_flags, currently unused]

Signed-off-by: Kir Kolyshkin <kir@openvz.org>